#304 refactor(reminder): Change alarm ID type from Long to Int in ControlAlarm
This commit updates the `ControlAlarm` interface and its implementation to use `Int` instead of `Long` for alarm identifiers. This aligns the ID type with the `PendingIntent.getBroadcast` request code requirement, removing the need for manual casting.

**Key Changes:**

*   **`ControlAlarm.kt`**:
    *   Changed the `idAlarm` and `alarmId` parameter types from `Long?` to `Int?` in `cancelAlarm`, `playRingtone`, and `stopRingtone`.
*   **`ControlAlarmImplementation.kt`**:
    *   Updated the implementation of `playRingtone`, `stopRingtone`, and `cancelAlarm` to accept `Int?` IDs.
    *   Removed the `.toInt()` conversion in `cancelAlarm` when creating the `PendingIntent`.

+++ app/src/main/kotlin/com/rahim/data/reminder/alarm/ControlAlarm.kt
@@ -4,7 +4,7 @@ import android.content.Context
 import android.media.Ringtone
 
 interface ControlAlarm {
-  fun cancelAlarm(context: Context, idAlarm: Long?)
-  fun playRingtone(context: Context, alarmId: Long?)
-  fun stopRingtone(context: Context, alarmId: Long?)
+  fun cancelAlarm(context: Context, idAlarm: Int?)
+  fun playRingtone(context: Context, alarmId: Int?)
+  fun stopRingtone(context: Context, alarmId: Int?)
 }
